+++ shiny3d/backend/gradio_shiny3d/__init__.py
@@ -1,0 +1,4 @@
+
+from .shiny3d import shiny3D
+
+__all__ = ['shiny3D']
