+++ shiny3d/demo/app.py
@@ -1,12 +1,11 @@
 
 import gradio as gr
-from shiny3dmodel import shiny3dmodel
+from gradio_shiny3d import shiny3D
 import os
 
 model_file = os.path.join(os.path.dirname(__file__), "Duck.glb")
 
-
 with gr.Blocks() as demo:
-    shiny3dmodel(interactive=True, value=model_file)
+    shiny3D(interactive=True, value=model_file)
 
 demo.launch()
