+++ shiny3d/frontend/example/Model3d.svelte
