+++ shiny3d/frontend/example/index.ts
