+++ shiny3d/frontend/icons/Minus.svelte
