+++ shiny3d/frontend/icons/Plus.svelte
