+++ shiny3d/frontend/package.json
@@ -1,16 +1,16 @@
 {
-  "name": "shiny3dmodel",
-  "version": "0.3.0-beta.4",
+  "name": "shiny3d",
+  "version": "0.3.0-beta.5",
   "description": "Gradio UI packages",
   "type": "module",
   "main": "./index.svelte",
   "author": "",
   "license": "ISC",
-  "private": true,
+  "private": false,
   "dependencies": {
     "@gradio/atoms": "0.2.0-beta.3",
     "@gradio/icons": "0.2.0-beta.0",
-    "@gradio/statustracker": "0.3.0-beta.4",
+    "@gradio/statustracker": "0.3.0-beta.5",
     "@gradio/upload": "0.3.0-beta.3",
     "@gradio/utils": "0.2.0-beta.3",
     "@types/babylon": "^6.16.6",
