+++ shiny3d/frontend/shared/utils.ts
