+++ shiny3d/frontend/static/Model3D.svelte
