+++ shiny3d/frontend/static/index.ts
