--- shiny3dmodel/backend/shiny3dmodel/__init__.py
@@ -1,4 +1,0 @@
-
-from .shiny3dmodel import shiny3dmodel
-
-__all__ = ['shiny3dmodel']
